* Added tests for `get_handler()` in handlers.__init__.py * Some changes in the other files to keep DRY. * Fixed bug uncovered by tests
